requirements.txt
apache-airflow==2.10.0
requests
kafka-python
pendulum